backend/requirement.txt
absl-py==1.4.0
appdirs==1.4.4
asgiref==3.5.2
astunparse==1.6.3
beautifulsoup4==4.12.2
cachetools==5.3.1
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
colorama==0.4.6
cryptography==41.0.1
defusedxml==0.7.1
Deprecated==1.2.14
distro==1.8.0
Django==4.1.1
django-braces==1.15.0
django-cors-headers==3.13.0
django-dotenv==1.4.2
django-oauth-toolkit==2.3.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
easyocr==1.7.0
et-xmlfile==1.1.0
filelock==3.12.2
flatbuffers==23.5.26
gast==0.4.0
google-auth==2.22.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.57.0
h5py==3.9.0
idna==3.4
imageio==2.31.1
imgkit==1.2.3
importlib-metadata==6.7.0
Jinja2==3.1.2
jwcrypto==1.5.0
keras==2.13.1
lazy_loader==0.2
libclang==16.0.6
lxml==4.9.2
Markdown==3.4.4
MarkupSafe==2.1.3
mpmath==1.3.0
mysqlclient==2.1.1
networkx==3.1
ninja==1.11.1
numpy==1.24.3
oauthlib==3.2.2
opencv-python-headless==4.7.0.72
openpyxl==3.1.2
opt-einsum==3.3.0
packaging==23.1
pandas==2.0.2
pdf2image==1.16.3
pdfkit==1.0.0
Pillow==9.5.0
protobuf==4.24.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pyclipper==1.3.0.post4
pycparser==2.21
pyee==8.2.2
PyJWT==2.5.0
PyPDF2==2.2.0
pyppeteer==1.0.2
python-bidi==0.4.2
python-dateutil==2.8.2
python-version==0.0.2
python3-openid==3.2.0
pytz==2022.2.1
PyWavelets==1.4.1
PyYAML==6.0
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
scikit-image==0.21.0
shapely==2.0.1
six==1.16.0
social-auth-app-django==5.2.0
social-auth-core==4.4.2
soupsieve==2.4.1
sqlparse==0.4.2
sympy==1.12
tabula-py==2.7.0
tensorboard==2.13.0
tensorboard-data-server==0.7.1
tensorflow==2.13.0
tensorflow-estimator==2.13.0
tensorflow-intel==2.13.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.3.0
tifffile==2023.4.12
torch==2.0.1
torchaudio==2.0.2
torchvision==0.15.2
tqdm==4.65.0
typing_extensions==4.5.0
tzdata==2022.2
urllib3==1.26.16
websockets==10.4
Werkzeug==2.3.7
wrapt==1.15.0
zipp==3.15.0